+++ app/src/main/java/com/example/controlenamao/model/Gasto.java
@@ -2,24 +2,24 @@ package com.example.controlenamao.model;
 
 public class Gasto {
 
-    private Integer id;
+    private Long id;
     public String name;
 
     public Gasto() {
     }
 
-    public Gasto(Integer id) {
+    public Gasto(Long id) {
         this.id = id;
     }
 
     public Gasto(String name) {
         this.name = name;
     }
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
     public String getName() {
